requirements.txt
aiogram==2.20
google-api-python-client==2.86.0
python-dotenv==0.21.0